Added some crude bit rate setting capability.

+++ src/samc21_can.h
@@ -78,7 +78,7 @@ public:
     * @return void
     */
     SAMC21_CAN(uint8_t _CS);
-    uint8_t begin(uint8_t idmodeset, uint8_t speedset, uint8_t clockset);
+    uint8_t begin(uint8_t idmodeset, uint32_t speedset, uint8_t clockset);
     uint8_t init_Mask(uint8_t num, uint8_t ext, uint32_t ulData);              // Initilize Mask(s)
     uint8_t init_Mask(uint8_t num, uint32_t ulData);                          // Initilize Mask(s)
     uint8_t init_Filt(uint8_t num, uint8_t ext, uint32_t ulData);              // Initilize Filter(s)
